requirements.txt
altair
pandas
streamlit
snowflake-snowpark-python
st_aggrid